BIG UPDATE TO THE 'search' COMMAND. click to read more!
Can now export icons AND featured icons! Can detect if an item has a featured icon or not. If it doesn't it grabs the icon.

The command can now also detect if the item you put in is a real item or not.

The command can now also ask if you want to tweet it.

It now saves the files as the item ID instead of just 'icon'.